Move EditorState to inlineEditing
Move inline editing related `Entry` functionality to `EntryDecorator`.

REDMINE-17645
