+++ entry_types/scrolled/package/src/frontend/Entry.js
@@ -1,10 +1,9 @@
-import React, {useEffect, useState, useCallback} from 'react';
+import React, {useState, useCallback} from 'react';
 
 import Chapter from "./Chapter";
 import MutedContext from './MutedContext';
 import ScrollToSectionContext from './ScrollToSectionContext';
-import {useEntryStructure, useEntryStateDispatch} from '../entryState';
-import {useEditorSelection} from './EditorState';
+import {useEntryStructure} from '../entryState';
 import {withInlineEditingDecorator} from './inlineEditing';
 import {usePostMessageListener} from './usePostMessageListener';
 
@@ -14,15 +13,13 @@ export default withInlineEditingDecorator('EntryDecorator', function Entry(props
   const [currentSectionIndex, setCurrentSectionIndexState] = useState(0);
 
   const [scrollTargetSectionIndex, setScrollTargetSectionIndex] = useState(null);
-  const {select} = useEditorSelection()
 
   const [muted, setMuted] = useState(true);
 
-  const dispatch = useEntryStateDispatch();
   const entryStructure = useEntryStructure();
 
   const setCurrentSectionIndex = useCallback(index => {
-    if (window.parent) {
+    if (window.parent !== window) {
       window.parent.postMessage({type: 'CHANGE_SECTION', payload: {index}}, window.location.origin);
     }
 
@@ -30,24 +27,12 @@ export default withInlineEditingDecorator('EntryDecorator', function Entry(props
   }, [setCurrentSectionIndexState]);
 
   const receiveMessage = useCallback(data => {
-    if (data.type === 'ACTION') {
-      dispatch(data.payload);
-    }
-    else if (data.type === 'SCROLL_TO_SECTION') {
+    if (data.type === 'SCROLL_TO_SECTION') {
       setScrollTargetSectionIndex(data.payload.index)
     }
-    else if (data.type === 'SELECT') {
-      select(data.payload);
-    }
-  }, [dispatch, select]);
+  }, []);
 
   usePostMessageListener(receiveMessage);
-
-  useEffect(() => {
-    if (window.parent !== window) {
-      window.parent.postMessage({type: 'READY'}, window.location.origin);
-    }
-  }, []);
 
   function scrollToSection(index) {
     if (index === 'next') {
